Split up SlackController into three classes
